backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
msgpack==1.1.0
packaging==24.1
psycopg2-binary==2.9.5
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.2
redis==4.5.5
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.2.3
